refactor(messenger): better code organization
