+++ integrations/messenger/src/index.ts
@@ -1,15 +1,11 @@
-import { IntegrationContext } from '@botpress/sdk'
 import { sentry as sentryHelpers } from '@botpress/sdk-addons'
-import { MessengerClient, MessengerTypes } from 'messaging-api-messenger'
 import queryString from 'query-string'
 import { idTag } from './const'
+import { handleMessage } from './misc/incoming-message'
+import { sendMessage } from './misc/outgoing-message'
+import { MessengerPayload } from './misc/types'
+import { formatGoogleMapLink, getCarouselMessage, getChoiceMessage, getMessengerClient } from './misc/utils'
 import * as bp from '.botpress'
-
-type Channels = bp.Integration['channels']
-type Messages = Channels[keyof Channels]['messages']
-type MessageHandler = Messages[keyof Messages]
-type MessageHandlerProps = Parameters<MessageHandler>[0]
-type IntegrationLogger = Parameters<bp.IntegrationProps['handler']>[0]['logger']
 
 const integration = new bp.Integration({
   register: async () => {},
@@ -190,208 +186,3 @@ export default sentryHelpers.wrapIntegration(integration, {
   environment: bp.secrets.SENTRY_ENVIRONMENT,
   release: bp.secrets.SENTRY_RELEASE,
 })
-
-type Carousel = bp.channels.channel.carousel.Carousel
-type Card = bp.channels.channel.card.Card
-type Choice = bp.channels.channel.choice.Choice
-type Dropdown = bp.channels.channel.dropdown.Dropdown
-type Location = bp.channels.channel.location.Location
-
-type MessengerAttachment = MessengerPostbackAttachment | MessengerSayAttachment | MessengerUrlAttachment
-
-type MessengerPostbackAttachment = {
-  type: 'postback'
-  title: string
-  payload: string
-}
-
-type MessengerSayAttachment = {
-  type: 'postback'
-  title: string
-  payload: string
-}
-
-type MessengerUrlAttachment = {
-  type: 'web_url'
-  title: string
-  url: string
-}
-
-function formatCardElement(payload: Card) {
-  const buttons: MessengerAttachment[] = []
-
-  payload.actions.forEach((action) => {
-    switch (action.action) {
-      case 'postback':
-        buttons.push({
-          type: 'postback',
-          title: action.label,
-          payload: `postback:${action.value}`,
-        })
-        break
-      case 'say':
-        buttons.push({
-          type: 'postback',
-          title: action.label,
-          payload: `say:${action.value}`,
-        })
-        break
-      case 'url':
-        buttons.push({
-          type: 'web_url',
-          title: action.label,
-          url: action.value,
-        })
-        break
-      default:
-        break
-    }
-  })
-  return {
-    title: payload.title,
-    image_url: payload.imageUrl,
-    subtitle: payload.subtitle,
-    buttons,
-  }
-}
-
-function getMessengerClient(ctx: bp.configuration.Configuration) {
-  return new MessengerClient({
-    accessToken: ctx.accessToken,
-    appSecret: ctx.appSecret,
-    appId: ctx.appId,
-  })
-}
-
-function getCarouselMessage(payload: Carousel): MessengerTypes.AttachmentMessage {
-  return {
-    attachment: {
-      type: 'template',
-      payload: {
-        templateType: 'generic',
-        elements: payload.items.map(formatCardElement),
-      },
-    },
-  }
-}
-
-function getChoiceMessage(payload: Choice | Dropdown): MessengerTypes.TextMessage {
-  if (!payload.options.length) {
-    return { text: payload.text }
-  }
-
-  if (payload.options.length > 13) {
-    return {
-      text: `${payload.text}\n\n${payload.options.map((o, idx) => `${idx + 1}. ${o.label}`).join('\n')}`,
-    }
-  }
-
-  return {
-    text: payload.text,
-    quickReplies: payload.options.map((option) => ({
-      contentType: 'text',
-      title: option.label,
-      payload: option.value,
-    })),
-  }
-}
-
-type SendMessageProps = Pick<MessageHandlerProps, 'ctx' | 'conversation' | 'ack'>
-
-async function sendMessage(
-  { ack, ctx, conversation }: SendMessageProps,
-  send: (client: MessengerClient, recipientId: string) => Promise<{ messageId: string }>
-) {
-  const messengerClient = getMessengerClient(ctx.configuration)
-  const recipientId = getRecipientId(conversation)
-  const { messageId } = await send(messengerClient, recipientId)
-  await ack({ tags: { [idTag]: messageId } })
-}
-
-export function getRecipientId(conversation: SendMessageProps['conversation']): string {
-  const recipientId = conversation.tags[idTag]
-
-  if (!recipientId) {
-    throw Error(`No recipient id found for user ${conversation.id}`)
-  }
-
-  return recipientId
-}
-
-async function handleMessage(
-  message: MessengerMessage,
-  {
-    client,
-    ctx,
-    logger,
-  }: { client: bp.Client; ctx: IntegrationContext<bp.configuration.Configuration>; logger: IntegrationLogger }
-) {
-  if (message?.message?.text) {
-    logger.forBot().debug('Received text message from Messenger:', message.message.text)
-    const { conversation } = await client.getOrCreateConversation({
-      channel: 'channel',
-      tags: {
-        [idTag]: message.sender.id,
-      },
-    })
-
-    const { user } = await client.getOrCreateUser({
-      tags: {
-        [idTag]: message.sender.id,
-      },
-    })
-
-    // TODO: do this for profile_pic as well, as of 13 NOV 2023 the url "https://platform-lookaside.fbsbx.com/platform/profilepic?eai=<eai>&psid=<psid>&width=<width>&ext=<ext>&hash=<hash>" is not working
-    if (!user.name) {
-      try {
-        const messengerClient = getMessengerClient(ctx.configuration)
-        const profile = await messengerClient.getUserProfile(message.sender.id, { fields: ['id', 'name'] })
-        logger.forBot().debug('Fetched latest Messenger user profile: ', profile)
-
-        await client.updateUser({ ...user, name: profile.name })
-      } catch (error) {
-        logger.forBot().error('Error while fetching user profile from Messenger:', error)
-      }
-    }
-
-    await client.createMessage({
-      tags: { [idTag]: message.message.mid },
-      type: 'text',
-      userId: user.id,
-      conversationId: conversation.id,
-      payload: { text: message.message.text },
-    })
-  }
-}
-
-function formatGoogleMapLink(payload: Location) {
-  return `https://www.google.com/maps/search/?api=1&query=${payload.latitude},${payload.longitude}`
-}
-
-type MessengerPayload = {
-  object: string
-  entry: MessengerEntry[]
-}
-
-type MessengerEntry = {
-  id: string
-  time: number
-  messaging: MessengerMessage[]
-}
-
-type MessengerMessage = {
-  sender: { id: string }
-  recipient: { id: string }
-  timestamp: number
-  message?: {
-    mid: string
-    text: string
-    quick_reply?: { payload: string }
-    attachments?: { type: string; payload: { url: string } }[]
-  }
-  postback?: {
-    mid: string
-    payload: string
-    title: string
-  }
-}
